feat:adding fastapi app with database connection and authentication

+++ TodoApp/routers/todos.py
@@ -1,16 +1,17 @@
 from fastapi import Depends, HTTPException, Path, APIRouter
-from models import Todos
+from models import Todos, User
 from database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session 
 from starlette import status
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 router=APIRouter()
 
 class TodoRequest(BaseModel):
-    title: str = Field(min_lenght=3)
-    description:str =  Field(min_lenght=3, max_lenght= 100)
+    title: str = Field(min_length=3)
+    description:str =  Field(min_length=3, max_length= 100)
     priority: int = Field(gt=0,lt=6)
     complete: bool
 
@@ -22,27 +23,36 @@ def get_db():
         db.close()
 
 db_dependency= Annotated[Session, Depends(get_db)]
+user_dependency=Annotated[dict,Depends(get_current_user)]
 
 @router.get("/")
-async def read_all(db:db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency,db:db_dependency):
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=404, detail='Authentication Failed')
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404,detail='Todo not found' )
 
 @router.post("/todo/add_data",status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user:user_dependency, db: db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=404, detail='Authentication Failed')
+    
+    todo_model = Todos(**todo_request.model_dump(),owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/update",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_id: int, todo_request: TodoRequest):
-    todo_model= db.query(Todos).filter(Todos.id==todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_id: int, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=404, detail='Authentication Failed')
+    todo_model= db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id == user.get('id')).first()
+
     if todo_model is None: 
         raise HTTPException(status_code=404, detail="Todo not found")
     
@@ -55,11 +65,13 @@ async def update_todo(db: db_dependency, todo_id: int, todo_request: TodoRequest
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency,todo_id :int = Path(gt=0)):
-    todo_model =db.query(Todos).filter(Todos.id==todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency,todo_id :int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=404, detail='Authentication Failed')
+    todo_model =db.query(Todos).filter(Todos.id==todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
